+++ src/recurrences/recurrences.controller.ts
@@ -7,6 +7,7 @@ import {
   Patch,
   Delete,
   Param,
+  BadRequestException,
 } from '@nestjs/common';
 import { RecurrencesService } from './recurrences.service';
 
@@ -16,16 +17,25 @@ export class RecurrencesController {
 
   @Get()
   findAll(@Headers('x-user-id') userId: string) {
+    if (!userId) {
+      throw new BadRequestException('x-user-id header is required');
+    }
     return this.recurrencesService.findAll(userId);
   }
 
   @Post()
   create(@Body() body: any, @Headers('x-user-id') userId: string) {
+    if (!userId) {
+      throw new BadRequestException('x-user-id header is required');
+    }
     return this.recurrencesService.create({ ...body, userId });
   }
 
   @Get(':id')
   findOne(@Param('id') id: string, @Headers('x-user-id') userId: string) {
+    if (!userId) {
+      throw new BadRequestException('x-user-id header is required');
+    }
     return this.recurrencesService.findOne(id, userId);
   }
 
@@ -35,11 +45,17 @@ export class RecurrencesController {
     @Body() body: any,
     @Headers('x-user-id') userId: string,
   ) {
+    if (!userId) {
+      throw new BadRequestException('x-user-id header is required');
+    }
     return this.recurrencesService.update(id, userId, body);
   }
 
   @Delete(':id')
   remove(@Param('id') id: string, @Headers('x-user-id') userId: string) {
+    if (!userId) {
+      throw new BadRequestException('x-user-id header is required');
+    }
     return this.recurrencesService.remove(id, userId);
   }
 }
